Sleep a little before exiting from subprocess
We ran into a issue where, during shutdown with processing job, the
program never exits. This is becuase the subprocess would shut down
before GRPC sends the results message to the manager. Thus the manger
waits for the results for ever. We should also close the results channel
on disconnect. Since we still want to receive the message, this sleep
is still necessary until I find a better way to wait for the message to
send.