BREAKING: Implement better locking
This commit change the way we lock jobs. We now build the Redis key for
a lock using an integer from 1 to the job's concurrency value. This
gives us a deterministic set of lock keys so that we can avoid scanning
the Redis key space. When we want to lock a job we randomly generate an
integer in the value above and loop until we find one or exhaust the
possibilities.

We treat 0 as a job that doesn't require a lock. And we treat -1 as a
failure to acquire a lock.

This job also changes some of the public interfaces to take pointers to
internal jobs.

The protobuf message format was updated.